Create controller and service interfaces
