+++ server/src/main/java/homeproject/webshop/webshop/controller/ProductController.java
@@ -1,29 +1,25 @@
 package homeproject.webshop.webshop.controller;
 
 import homeproject.webshop.webshop.domain.Product;
-import homeproject.webshop.webshop.service.ProductService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@CrossOrigin
-@RestController
-@RequestMapping("/product")
-public class ProductController {
+public interface ProductController {
 
-    @Autowired
-    private ProductService productService;
+    /**
+     * Add a product
+     *
+     * @param product Added product
+     * @return ResponseEntity<Product>
+     */
+    ResponseEntity<Product> addProduct(Product product);
 
-    @PostMapping("/addProduct")
-    public ResponseEntity<Product> addProduct(@RequestBody Product product){
-        return ResponseEntity.ok(productService.addProduct(product));
-    }
-
-    @GetMapping("/getAllProducts")
-    public ResponseEntity<List<Product>> getAllProducts(){
-        return ResponseEntity.ok(productService.getAllProducts());
-    }
+    /**
+     * Get all products
+     *
+     * @return ResponseEntity<List<Product>>
+     */
+    ResponseEntity<List<Product>> getAllProducts();
 
 }
