+++ server/src/main/java/homeproject/webshop/webshop/controller/ProductControllerImpl.java
@@ -1,0 +1,31 @@
+package homeproject.webshop.webshop.controller;
+
+import homeproject.webshop.webshop.domain.Product;
+import homeproject.webshop.webshop.service.ProductServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@CrossOrigin
+@RestController
+@RequestMapping("/product")
+public class ProductControllerImpl implements ProductController {
+
+    @Autowired
+    private ProductServiceImpl productService;
+
+    @Override
+    @PostMapping("/addProduct")
+    public ResponseEntity<Product> addProduct(@RequestBody Product product){
+        return ResponseEntity.ok(productService.addProduct(product));
+    }
+
+    @Override
+    @GetMapping("/getAllProducts")
+    public ResponseEntity<List<Product>> getAllProducts(){
+        return ResponseEntity.ok(productService.getAllProducts());
+    }
+
+}
