+++ server/src/main/java/homeproject/webshop/webshop/controller/WebShopUserController.java
@@ -1,38 +1,33 @@
 package homeproject.webshop.webshop.controller;
 
 import homeproject.webshop.webshop.domain.WebShopUser;
-import homeproject.webshop.webshop.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@RestController
-@CrossOrigin
-@RequestMapping("/user")
-public class WebShopUserController {
+public interface WebShopUserController {
 
-    @Autowired
-    private UserService userService;
+    /**
+     * Add a user
+     *
+     * @param user Added user
+     * @return ResponseEntity<WebShopUser>
+     */
+    ResponseEntity<WebShopUser> addUser(WebShopUser user);
 
-    @PostMapping("/addUser")
-    public ResponseEntity<WebShopUser> addUser(@RequestBody WebShopUser user){
-        return ResponseEntity.ok(userService.addUser(user));
-    }
+    /**
+     * Get all users
+     *
+     * @return ResponseEntity<List<WebShopUser>>
+     */
+    ResponseEntity<List<WebShopUser>> getAllUsers();
 
-    @GetMapping("/getAllUsers")
-    public ResponseEntity<List<WebShopUser>> getAllUsers(){
-        return ResponseEntity.ok(userService.getAllUser());
-    }
+    /**
+     * Delete a user
+     *
+     * @param id Id of the user
+     * @return ResponseEntity<WebShopUser>
+     */
+    ResponseEntity<WebShopUser> deleteUser(Long id);
 
-    @GetMapping("/deleteUser/{id}")
-    public ResponseEntity<WebShopUser> deleteUser(@PathVariable Long id){
-        return ResponseEntity.ok(userService.deleteUser(id));
-    }
-
-    @PostMapping("/login")
-    public boolean login(@RequestBody WebShopUser user) {
-        return userService.isUserExists(user);
-    }
 }
