+++ server/src/main/java/homeproject/webshop/webshop/controller/WebShopUserControllerImpl.java
@@ -1,0 +1,42 @@
+package homeproject.webshop.webshop.controller;
+
+import homeproject.webshop.webshop.domain.WebShopUser;
+import homeproject.webshop.webshop.service.UserServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@CrossOrigin
+@RequestMapping("/user")
+public class WebShopUserControllerImpl implements WebShopUserController{
+
+    @Autowired
+    private UserServiceImpl userService;
+
+    @Override
+    @PostMapping("/addUser")
+    public ResponseEntity<WebShopUser> addUser(@RequestBody WebShopUser user){
+        return ResponseEntity.ok(userService.addUser(user));
+    }
+
+    @Override
+    @GetMapping("/getAllUsers")
+    public ResponseEntity<List<WebShopUser>> getAllUsers(){
+        return ResponseEntity.ok(userService.getAllUser());
+    }
+
+    @Override
+    @GetMapping("/deleteUser/{id}")
+    public ResponseEntity<WebShopUser> deleteUser(@PathVariable Long id){
+        return ResponseEntity.ok(userService.deleteUser(id));
+    }
+
+    //TODO auth
+    @PostMapping("/login")
+    public boolean login(@RequestBody WebShopUser user) {
+        return userService.isUserExists(user);
+    }
+}
