+++ server/src/main/java/homeproject/webshop/webshop/service/ProductService.java
@@ -1,26 +1,24 @@
 package homeproject.webshop.webshop.service;
 
 import homeproject.webshop.webshop.domain.Product;
-import homeproject.webshop.webshop.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-public class ProductService {
+public interface ProductService {
 
-    @Autowired
-    private ProductRepository productRepository;
+    /**
+     * Add a product
+     *
+     * @param product Added product
+     * @return Product
+     */
+    Product addProduct(Product product);
 
-    //TODO
-    public Product addProduct(Product product){
-        return productRepository.save(product);
-    }
-
-    //TODO //????
-    public List<Product> getAllProducts(){
-        return productRepository.findAll();
-    }
+    /**
+     * Get all products
+     *
+     * @return List<Products>
+     */
+    List<Product> getAllProducts();
 
 }
