+++ server/src/main/java/homeproject/webshop/webshop/service/ProductServiceImpl.java
@@ -1,0 +1,26 @@
+package homeproject.webshop.webshop.service;
+
+import homeproject.webshop.webshop.domain.Product;
+import homeproject.webshop.webshop.repository.ProductRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ProductServiceImpl implements ProductService{
+
+    @Autowired
+    private ProductRepository productRepository;
+
+    @Override
+    public Product addProduct(Product product){
+        return productRepository.save(product);
+    }
+
+    @Override
+    public List<Product> getAllProducts(){
+        return productRepository.findAll();
+    }
+
+}
